Refactor creds as a full Jenkins server object
Instead of having sets of credentials config nodes and repeating
the URL in every process nodes, this creates a config object
describing a server (creds + Url) that also holds the API object.
