+++ nodes/build.html
@@ -4,8 +4,7 @@
         category: 'jenkins',
         color: '#C7E9C0',
         defaults: {
-            creds: { type: "jenkins-credentials", required: false },
-            url: { value: "" },
+            server: { type: "jenkins-server", required: false },
             job: { value: "", validate: RED.validators.typedInput("jobType") },
             jobType: { value: "" },
             buildNumber: { value: "", validate: RED.validators.typedInput("buildNumberType") },
@@ -39,12 +38,8 @@
 
 <script type="text/x-red" data-template-name="jenkins-build">
     <div class="form-row">
-        <label for="node-input-creds"><i class="fa fa-key"></i> <span data-i18n="job.label.config.credentials"></span></label>
-        <input type="text" id="node-input-creds">
-    </div>
-    <div class="form-row">
-        <label for="node-input-url"><i class="icon-globe"></i> <span data-i18n="job.label.config.url"></span></label>
-        <input type="text" id="node-input-url" placeholder="URL">
+        <label for="node-input-server"><i class="fa fa-key"></i> <span data-i18n="job.label.config.server"></span></label>
+        <input type="text" id="node-input-server">
     </div>
     <div class="form-row">
         <label for="node-input-job"><i class="icon-cog"></i> <span data-i18n="job.label.config.job"></span></label>
