+++ nodes/config-server.html
@@ -1,0 +1,41 @@
+<script type="text/x-red" data-template-name="jenkins-server">
+    <div class="form-row">
+          <label for="node-config-input-name"><i class="fa fa-tag"></i> <span data-i18n="jenkins-server.label.config.name"></span></label>
+          <input type="text" id="node-config-input-name">
+    </div>
+
+    <div class="form-row">
+      <label for="node-config-input-username"><i class="fa fa-user"></i> <span data-i18n="jenkins-server.label.config.username"></span></label>
+      <input type="text" id="node-config-input-username">
+    </div>
+
+    <div class="form-row">
+      <label for="node-config-input-password"><i class="fa fa-key"></i> <span data-i18n="jenkins-server.label.config.password"></span></label>
+      <input type="password" id="node-config-input-password">
+    </div>
+
+    <div class="form-row">
+      <label for="node-config-input-url"><i class="fa fa-globe"></i> <span data-i18n="jenkins-server.label.config.url"></span></label>
+      <input type="text" id="node-config-input-url" placeholder="https://your-jenkins-server">
+    </div>
+</script>
+
+<script type="text/javascript">
+    RED.nodes.registerType('jenkins-server',{
+      defaults: {
+        name: {value:"", required: true},
+        url: {type: "string", required: true}
+      },
+      category: 'config',
+      credentials: {
+        username: {type: "string", required: true},
+        password: {type: "password", required: true},
+      },
+      label: function() {
+        return this.name || this.credentials.username + "@" + this.url;
+      },
+      color:"#787878",
+      icon: "jenkins.png",
+      exportable: false
+    });
+</script>
