--- nodes/credentials.html
@@ -1,35 +1,0 @@
-<script type="text/x-red" data-template-name="jenkins-credentials">
-    <div class="form-row">
-          <label for="node-config-input-name"><i class="fa fa-tag"></i> <span data-i18n="jenkins-credentials.label.config.name"></span></label>
-          <input type="text" id="node-config-input-name">
-    </div>
-
-    <div class="form-row">
-      <label for="node-config-input-username"><i class="fa fa-user"></i> <span data-i18n="jenkins-credentials.label.config.username"></span></label>
-      <input type="text" id="node-config-input-username">
-    </div>
-
-    <div class="form-row">
-      <label for="node-config-input-password"><i class="fa fa-key"></i> <span data-i18n="jenkins-credentials.label.config.password"></span></label>
-      <input type="password" id="node-config-input-password">
-    </div>
-</script>
-
-<script type="text/javascript">
-    RED.nodes.registerType('jenkins-credentials',{
-      defaults: {
-        name: {value:"", required: true},
-      },
-      category: 'config',
-      credentials: {
-        username: {type: "string", required: true},
-        password: {type: "password", required: true}
-      },
-      label: function() {
-        return this.name || this.user;
-      },
-      color:"#787878",
-      icon: "jenkins.png",
-      exportable: false
-    });
-</script>
