+++ nodes/job.js
@@ -1,6 +1,3 @@
-URI = require('urijs');
-util = require('util');
-
 module.exports = function(RED) {
     "use strict";
 
@@ -11,32 +8,15 @@ module.exports = function(RED) {
     function job(config) {
         RED.nodes.createNode(this, config);
 
-        this.creds = config.creds
-        this.url = config.url;
-
+        this.server = config.server
         this.job = config.job;
         this.jobType = config.jobType;
         this.parameters = config.parameters;
 
         let node = this;
         this.on('input', function(msg) {
-            let jenkins = require('jenkins');
-            let jenkinsInstance = null
+            let jenkins_server = RED.nodes.getNode(node.server);
 
-            {
-                let url = new URI(node.url)
-                  .username(RED.nodes.getNode(node.creds).credentials.username)
-                  .password(RED.nodes.getNode(node.creds).credentials.password);
-
-                let jenkinsConfig = {
-                    baseUrl: url.toString(),
-                    crumbIssuer: true,
-                };
-
-                jenkinsInstance = jenkins(jenkinsConfig);
-            }
-
-            this.log("Running job " + this.job + this.parameters)
             function getValue(type, value) {
                 if (type === 'msg') {
                     return RED.util.getMessageProperty(msg, value);
@@ -49,7 +29,7 @@ module.exports = function(RED) {
             }
 
             let job = getValue(node.jobType, node.job)
-            jenkinsInstance.job.build({name: job, parameters: this.parameters}, function(err, data) {
+            jenkins_server.api.job.build({name: job, parameters: this.parameters}, function(err, data) {
                 if (err) {
                     node.status({ 
                                   fill: "red",
