+++ nodes/locales/en-US/build.json
@@ -2,7 +2,7 @@
   "job": {
     "label": {
       "config": {
-        "credentials": "Jenkins ID",
+        "server": "Jenkins server",
         "url": "URL",
         "job": "Job",
         "buildNumber": "Build number"
