+++ nodes/locales/en-US/config-server.json
@@ -1,5 +1,5 @@
 {
-  "jenkins-credentials": {
+  "jenkins-server": {
     "label": {
       "config": {
         "name": "Jenkins ID",
