+++ nodes/locales/en-US/job.json
@@ -2,7 +2,7 @@
   "job": {
     "label": {
       "config": {
-        "credentials": "Jenkins ID",
+        "server": "Jenkins server",
         "url": "URL",
         "job": "Job",
         "parameters": "Parameters"
