+++ nodes/status.html
@@ -4,7 +4,7 @@
         category: 'jenkins',
         color: '#168BB9',
         defaults: {
-            creds: { type: "jenkins-credentials", required: false },
+            creds: { type: "jenkins-server", required: false },
             url: { value: "" }
         },
         inputs:1,
